Update conf for new sensor

+++ main/adc_readings.h
@@ -6,6 +6,6 @@
 
 // void get_battery_percentage(void *pvParameters);
 int get_battery_percentage(void);
-double get_solar_irradiance(void);
+float get_solar_irradiance(void);
 
 #endif //PROJ_ADC_READINGS_H_
